+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Added config options to prevent the mod from disabling attack collisions with/from specific characters
 
+* Added config option to prevent the mod from disabling projectile collisions of projectiles that can stick to characters (loader grapple, engi spider mines, etc)
+
 * Fixed player projectiles still impacting allies if a player changes to another team.
 
 **1.0.1 Changes:**